Fix TocSequenceEventImpl to get exactly the behavior from before PR #7
This does not change the final TOC, however it does change the
intermediary ones, more precisely when some elements that are referred
to by a toc-entry don’t have a volume number assigned to them
yet (i.e. in the first VolumeProvider iteration and in the next
iterations when not all content has been fit into the reserved volumes
yet). And this in turn can result in a difference in the final result
because the volume breaking algorithm is so sensitive.