Remove EvenSizeVolumeSplitter code
The reason is explained in issue
https://github.com/mtmse/dotify.formatter.impl/issues/5. While an
algorithm that favours solutions with more even volumes would be
useful to have, EvenSizeVolumeSplitter (despite its name) does not
actually accomplish this. Moreover, it has many problems,
deteriorating the volume breaking more than improving it. Until we
find a way to implement it properly, the whole feature is dropped.